Refactor configuration into seperate config file
