+++ tools/__tasks__/screenshot/check-network.js
@@ -1,14 +1,13 @@
 const tcpp = require('tcp-ping');
 const pify = require('pify');
 
-const domain = 'localhost';
-const port = 9000;
+const {port, domain} = require('./config');
 
 module.exports = {
     description: `Probing ${domain} on port ${port}...`,
     task: () => pify(tcpp.probe, {multiArgs: true})(domain, port).then((result) => {
         if (!result[0]) {
-            throw new Error('Cannot reach the network - is your server running?');
+            throw new Error(`Cannot reach ${domain}:${port} - is your server running?`);
         }
     })
 };
