+++ tools/__tasks__/screenshot/index.js
@@ -1,5 +1,5 @@
 module.exports = {
-    description: 'Oke Doke, taking some screenshots for you... This may take a while',
+    description: 'Ok, taking some screenshots for you... This may take a while ⏳',
     task: [
         require('./check-network'),
         require('./screenshot')
